Update FlexRegBuilder.java
Added functionality to the multiplication button and save button. The multiplication button now displays the input data and inserts a multbean object into the arraylist. The save button now allows the user to select which file they want to save to by clicking on a loaded file in the left panel.