Update FlexRegBuilder.java
Altered the name of the delete button to "delete file", fixed bug with delete button where it could delete test cases instead.
Removed Address Button since it has no current purpose.
Made it so the add button will add a file to the right panel and the arraylist that holds the objects in the right list.
**NOTE** the XMLDecoder currently only takes the first object in the xml file when adding