Fixed a mistake with "Add" button error checking
Originally, pressing the Add button will almost always cause "Error: Select a file to add" even when a file is selected. This is because it seems that the only way to stop reading from an XML file is to throw an ArrayIndexOutOfBoundsException. I had to fix this by checking if a file isn't selected a different way; seeing if the index is -1.